# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Smart Internship Matching Portal REST API",
    "version": "1.0.0",
    "description": "Enterprise API documentation mapping Students, Recruiters, and Admins alongside matching engines."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Register a new user (Student or Recruiter)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "role", "name"],
                "properties": {
                  "email": { "type": "string", "example": "student@example.com" },
                  "password": { "type": "string", "example": "Password@123" },
                  "role": { "type": "string", "enum": ["student", "recruiter"], "example": "student" },
                  "name": { "type": "string", "example": "Alex Rivera" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created. OTP sent." }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Log in and obtain JWT tokens",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "example": "student@example.com" },
                  "password": { "type": "string", "example": "Password@123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Session tokens returned." }
        }
      }
    },
...
  }
}